Semi Transparent lines in CompareSetPlots
Added an rgb value with 0.5 transparency to the red/blue lines in the CompareSetPlots feature -- this makes overlaps easier to notice, so that perfect matches are harder to confuse with complete mismatches.  Also changed the par() to revert to 1x1, since that ought to be the default and the oldpar thing wasn't working.  dev.off() clears the plots, so that's no fun; need to fix that later, but this is better for now.